Renamed IntakeNote to just Intake.
