+++ src/main/java/frc/robot/RobotContainer.java
@@ -24,8 +24,8 @@ import frc.lib.utils.AllianceFlipUtil;
 import frc.robot.commands.drive.DriveWithDpad;
 import frc.robot.commands.drive.DriveWithJoysticks;
 import frc.robot.commands.leds.LedRainbowCMD;
-import frc.robot.subsystems.IntakeNote.IntakeNote;
-import frc.robot.subsystems.IntakeNote.IntakeNoteIO;
+import frc.robot.subsystems.Intake.IntakeNote;
+import frc.robot.subsystems.Intake.IntakeNoteIO;
 import frc.robot.subsystems.drive.Drive;
 import frc.robot.subsystems.drive.ModuleIO;
 import frc.robot.subsystems.drive.ModuleIOSim;
