+++ src/main/java/frc/robot/subsystems/Intake/IntakeNote.java
@@ -2,11 +2,13 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.subsystems.IntakeNote;
+package frc.robot.subsystems.Intake;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import frc.robot.subsystems.IntakeNote.IntakeNoteIOInputsAutoLogged;
+
 import org.littletonrobotics.junction.Logger;
 
 public class IntakeNote extends SubsystemBase {
